+++ Sources/ClingsCLI/Clings.swift
@@ -27,7 +27,7 @@ struct Clings: AsyncParsableCommand {
         For more information on a specific command, run:
           clings <command> --help
         """,
-        version: "0.3.3",
+        version: "0.4.0",
         subcommands: [
             // List views
             TodayCommand.self,